Rename utils to common, and realise that the Utils file in test/integration is misplaced. Dissolve it by duplicating 8 lines of code rather than finding it a new home.
