+++ reasoner/common/Planner.java
@@ -16,7 +16,7 @@
  *
  */
 
-package com.vaticle.typedb.core.reasoner.utils;
+package com.vaticle.typedb.core.reasoner.common;
 
 import com.vaticle.typedb.core.common.exception.TypeDBException;
 import com.vaticle.typedb.core.logic.resolvable.Concludable;
