+++ reasoner/reactive/AbstractReactive.java
@@ -21,7 +21,7 @@ package com.vaticle.typedb.core.reasoner.reactive;
 import com.vaticle.typedb.core.reasoner.reactive.common.ReactiveActions;
 import com.vaticle.typedb.core.common.iterator.FunctionalIterator;
 import com.vaticle.typedb.core.reasoner.reactive.common.Operator;
-import com.vaticle.typedb.core.reasoner.utils.Tracer;
+import com.vaticle.typedb.core.reasoner.common.Tracer;
 
 import java.util.function.Function;
