+++ reasoner/reactive/Output.java
@@ -18,7 +18,7 @@
 
 package com.vaticle.typedb.core.reasoner.reactive;
 
-import com.vaticle.typedb.core.reasoner.utils.Tracer;
+import com.vaticle.typedb.core.reasoner.common.Tracer;
 
 /**
  * Governs an output from a reactiveBlock
