--- test/integration/reasoner/utils/Util.java
@@ -1,40 +1,0 @@
-/*
- * Copyright (C) 2021 Vaticle
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- *
- */
-
-package com.vaticle.typedb.core.reasoner.utils;
-
-import com.vaticle.typedb.core.logic.LogicManager;
-import com.vaticle.typedb.core.pattern.Conjunction;
-import com.vaticle.typedb.core.pattern.Disjunction;
-import com.vaticle.typeql.lang.TypeQL;
-
-public class Util {
-
-    public static Disjunction resolvedDisjunction(String query, LogicManager logicMgr) {
-        Disjunction disjunction = Disjunction.create(TypeQL.parsePattern(query).asConjunction().normalise());
-        logicMgr.typeInference().applyCombination(disjunction);
-        return disjunction;
-    }
-
-    public static Conjunction resolvedConjunction(String query, LogicManager logicMgr) {
-        Disjunction disjunction = resolvedDisjunction(query, logicMgr);
-        assert disjunction.conjunctions().size() == 1;
-        return disjunction.conjunctions().get(0);
-    }
-
-}
